VCI: don't send "qos commit" when nothing changed
With deferred or non-existent interfaces, QoS VCI can be presented with
a new JSON config, but due to the interfaces not having an ifindex, no
configuration commands are sent down to the dataplane.

In these circumstances we shouldn't send the final "qos commit" command.

Also added logging of commands as they are sent to vplaned.

Part of VRVDR-49148