VCI: add support for bonded interfaces
Because QoS policies are augmented to bonded interfaces in a yang
file that belongs to bonded interfaces and not QoS, we need to add
vyatta-interfaces-bonding-qos-v1 to the list of yang files that
QoS VCI is interested in.

However as the bonding yang file has a different namespace from the
standard QoS yang files which leads to some additional code paths
when trying to retrieve the QoS policy name that's attached to a
bonded interface.

See the new test-data in tests/test_interface.py for examples of these
namespace differences.

Part of VRVDR-49148