VCI: add support for deferred/vhost interfaces
Vhost interfaces are one type of interface that can be deferred. That
is, QoS VCI can be presented with a valid QoS config for a vhost
interface that doesn't yet exist on the system.   Vhost interfaces only
come into existance when the guest VM that they are connected to are
active and running on the host VNF system.

Deferred interfaces are identified within QoS VCI by having a ifindex
of None.

To be told about deferred interfaces being "plugged into" the system,
QoS VCI now listens to "interface-state" notifications from ifmgr.

These "interface-state" notifications only tell QoS VCI about the
interface being created/deleted.  It is not presented with any
configuration data related to the interface.

Because of this QoS, now keeps two, possibly different, configuration
files on disk.  The "requested" config file holds all the QoS config
data that was presented to it at the time of the last commit.  While
processing this data, QoS VCI may have identified some deferred
interfaces for which it was unable to send config commands to the
dataplane.  QoS VCI creates the "actioned" config file by removing
any deferred interface configuration from the "requested" config file.

When QoS VCI receives an "interface-state" notification for a
deferred interface, it compares the "actioned" config against the
"requested" config to see if the new interface has any qos config
waiting to be sent to the dataplane

As QoS policies are attached to vhost interfaces by and augment in
vyatta-interfaces-vhost-qos-v1, we have some additional yang namespace
and element naming problems to handle.

See the new vhost test-data for details of the namespace/element naming
changes.

Fixes: VRVDR-49148