VCI: use the FWHelper.pm script to build NPF rules
Use NPF's FWHelper.pm perl script to build NPF rules.

Prior to this commit QoS VCI had its Rule.py module that provides a
limited emulation of the FWHelper.pm script.  Whilst it covered many
of the NPF rules, it didn't cover DPI and a few of the more obscure
rules.

The new Rule.py module spawns a subprocess to call a perl script that
creates a configd session and then calls the FWHelper.pm script to
generate the necessary NFP rule command.

Fixes: VRVDR-48986