VCI: don't modify ingress-maps in-place
Prior to this commit ingress-map could essentially be modified
in-place.  After this commit, ingress-maps are modified using the same
method as policies, they are first deleted, and then re-created.

As part of this, before an ingress-map is deleted any bindings to
ports or vlans must be broken.  To do this the generation of the
ingress-map binding commands have been moved from the provisioner's
_create_interface/_delete_interface methods to the ingress-map's
commands and delete_cmd methods.

This has a couple of knock-on effects.  When ingress-maps are bound to
a port or vlan, we need the interface name for the ctrl.store method
as the ingress-map binding is stored against a particular interface.
This means that the _obj_create/_obj_delete methods for global objects
must now generate a list of three tuples (path, cmd, ifname) rather
than a single (path, cmd) command.  This means that the unit-tests for
the action-group, mark-map and ingress-map classes also need updated.

VRVDR-49756