Fix the "show .... <if-name> class" output
When the NPF team rejigged the JSON generated by the vyatta-dataplane
when queried about NPF rules with rprocs, they broke the:
"show queuing <if-name> class" and "show policy qos <if-name> class"
command by moving some of the data that these commands displayed into
new JSON elements.

This commit fishes around in these new JSON elements to extract the
data that used to be elsewhere.

Fixes: VRVDR-49194