Fix hardware or sofware support based on config
We can currently only support queue limits in bytes in hardware
and only support queue limits in packets in software (DPDK).

Rather than using the state of hw forwarding in the dataplane to choose
whether the port should be set up to do hardware or software QoS,
and then failing to configure the port correctly if the queue limit
type is incompatible, use this setting (via a new parameter on the
qos port command) to force the port to be configured for one or the other.

VRVDR-49973