VCI: emulate FWHelper.pm's build_rule function
In an attempt to reduce maintanence overhead, we previously decided
to try to use NPF's FWHelper.pm perl script to process the QoS
match rule configuration and return the constructed NPF rule command.

This approach has been shown not to work as to access the FWHelper.pm
script it is necessary to spawn a sub-process to get from the python
interpreter to the perl interpreter.  Unfortunately the sub-process
doesn't get access to the candidate configuration that the python
QOS VCI process has been presented with.   The sub-process can only
see the current running configuration.  This means that the NPF
rule command returned by the sub-process doesn't contain any new match
criteria that are present in the candidate configuration.

The fix is to emulate FWHelper.pm's build_rule function fully in
python.

Fixes: VRVDR-49700