use interface name rather than ifindex in cfg cmds
The vplane controller's ephemeral interface support should
handle config commands applied to an interface wrt to ordering
versus the interface netlink create. If the interface is deleted
and recreated with the same name then it will apply that config
to the recreated interface after the interface is created in the
dataplane.

This new infra relies on config commands using an interface name
rather than an ifindex, so convert QoS to use the interface name
in all its config commands. For "global" objects, ie. ones not
interface specific, the special value of 0 (invalid ifindex) is
changed to the string "global-object_cmd" which is longer than
the linux IFNAMSIZ and therefore will never be confused with a
valid linux interface name.

VRVDR-50255