VCI: log unhandled exceptions in system log
When the QoS VCI process crashes due to an unhandled exception a fairly
meaningless stacktrace originating in libvci is logged in the system
log. The logged stacktrace contained no useful information in
determining the cause of the QoS VCI process crash.

To improve the debuggability of QoS VCI, catch these previously
unhandled at the highest level and log meaningful traceback information
to the system log.

Part of: VRVDR-50316