VCI: remove redundant deferred config code
Moving the responsibility of dealing with deferred interfaces (for
example vhost interfaces) from the QoS VCI code down to the dataplane
means that QoS VCI no longer needs to keep a requested and an actioned
config file.

This commit removes the code that is no longer required.

Part of: VRVDR-49636