VCI: op-mode state not fully populated
The op-mode state wasn't being fully populated, a traffic-class would
go missing when the reverse_map for that traffic-class wasn't complete
because the attempted access to the reverse_map would case a KeyError
exception, and the queue_out object wouldn't be appended to the
tc_queues_out list.

Moved the append of the tc_queues_out list out of the key/exception
block.

Fixes: VRVDR-50413