VCI: catch IndexError exception in qos_op_mode.py
Using an "if ... is not None:" to check that an entry of a list of
dictionaries is populated doesn't work, it just throws an IndexError
exception.

Recoded the function to use try/except blocks.

During testing also found a try/except block where both an IndexError
and KeyError exception could be raised.

Fixes: VRVDR-50473