Remove mark-map on trunk check
Remove the validation check that doesn't allow a mark-map
to be used in a policy attached as a trunk policy. This
check wasn't run when the attachment is via a switchport. So
remove it from routed port attachment as well to keep things
consistent and allow remarking of PCP for VLANs that don't
have a more specific subport policy.

VRVDR-50342