VCI: detaching a qos policy needs a delete/create
When a qos policy is detached from an interface, we can't just add
the "new version" of the interface to the _if_update list because
the "new" interface doesn't know the name of the policy being deleted.

So we need to add the "old" version of the interface to the _if_delete
list, and the "new" version of the interface to the _if_create list.

Fixes: VRVDR-50529