VCI: ratelimit period does not override default
When a policer's ratelimit command does not have a "pps" suffix
the policer's period value should be passed down to the NPF code.
Prior to this commit, it was not happening, as the ratelimit period
was fixed at 1000 milliseconds.

Fixes: VRVDR-50820
(cherry picked from commit 0f87719ba11397baf8c13f525aeb7a7fcd0f646e)