VCI: stop QoS VCI crashing with incomplete config
If the user configured a profile with a dscp-group map to assigned
a particular dscp-group into a pipe-queue, but did not specify which
traffic-class that the pipe-queue should be assigned to, then QoS VCI
would throw an unhandled exception that would result in an error being
reported back at commit time.

To make dscp-group maps behave in the same manner as the other pcp and
dscp maps, this incomplete configuration is ignored.

Fixes: VRVDR-50967
(cherry picked from commit 7e06455821bf87add15a7a16e73e531964894624)