Don't send all action group json in one field
The dataplane optimised-show output for action groups has changed
to a nested arrangement of JSON objects rather than a single
string which could be put into the yang action-group field, which
is of string type. Returning the nested JSON is not acceptable
to netconf.

As the policer stats now have their own entries, the only relevant
thing to include here is the name of the action group.

VRVDR-51108