NPF rule "bytes" and "packets" are 64-bit counters
The NPF rule "bytes" and "packet" leafs are defined as uint64 in the
QoS op-mode yang, but the qos-op-mode.pl and the qos_op_mode.py scripts
present the values as JSON integers.  64-bit integer values should be
presented as strings, i.e. the numbers should be enclosed in double
quotes.

Fixes: VRVDR-52905