VCI: allow a class profile to be a global profile
The VCI qclass command method didn't check for a global profile
index if the local profile index returned None.

This meant that the vyatta qos dataplane config command:

  "qos <if-name> pipe <subport-id> <pipe-id> <profile-id>"

was sent with "None" rather than a numerical <profile-id>.

This would cause the QoS config code in the dataplane to log the
following error:

  DATAPLANE: profile id not a number: None

and reject the new config.  If the dataplane crashed for any
reason, it would then constantly restart due to the invalid
vyatta qos dataplane config command

Also updated VCI unit-tests to check for this condition.

Fixes: VRVDR-53599 and DAN-219