VCI: stop QoS VCI crashing when dataplane is down
If the dataplane is down (or in the process of restarting) and unable
to respond to op-mode requests from the vplaned-controller, then the
Controller context-manager will raise a ZMQ Again exception before the
yang_state variable has been initialised.  Once the generic exception
handler logged the "Unhandled exception" the State.get method would
try to return using the uninitialised yang_state variable causing QoS
VCI to crash.

Rearranged the State.get method to handle the ZMQ Again exception and
to not use the uninitialised yang_state variable.

Fixes: VRVDR-53805