QoS: Handle LAG membership change notifications
When LAG membership changes, the LAG component sends a VCI notification
to the QoS component so it can attach/detach the LAG QoS policy on the
physical interface accordingly.

A new function to handle the VCI notifications from the LAG component
is registered by the QoS component.
The Provisioner object distinguishes between a new QoS configuration
from configd and a LAG membership change VCI notification from the
LAG component. In the notification case, Provisioner instantiates
two QosConfig objects as usual (old and new config) from the current
QoS configuration, without interface, and updates one of the two
objects with the LAG member interface reported in the notification
(e.g. if the notification reports that a LAG has a new member, that
interface will be added to the QosConfig object that represent the new
configuration). The difference between the old and the new
QosConfig will make the appropriate QoS configuration changes
to be applied in the LAG member.

There is a special case where QoS and LAG configuration changes
are being applied in the same CLI commit. I such cases, configd
will first apply the QoS configuration and then the LAG
configuration. That means that the QoS configuration may be
incomplete until the LAG changes are applied and the LAG
component sends the VCI notification, so the QoS component can
have an up-to-date information regarding LAG membership. For
example:

A QoS policy is attached to a LAG, a LAG member is removed and
another policy is attached to the dataplane port (former LAG
member), all in the same CLI commit. The QoS configuration will
be applied first by configd, and Provisioner will have a QoS
config that contains a policy being attached to a LAG and
another policy that is being attached to a dataplane port, that
is also a LAG member (QoS hasn't received yet the LAG membership
update). In such case, the LAG QoS policy must be favoured
over the dataplane policy configuration. The policy will be
attached to all LAG members, including the one that is being
removed, and the dataplane policy will be ignored for now.
Once the LAG membership change notification is received, the
LAG policy will be detached from the dataplane port and the
dataplane QoS policy that was stored in the configuration
JSON file will be applied.

VRVDR-49815