Copy the traffic_type from Rules into GPCTable
Both the GPCTable and the Rules protobuf have a traffic_type field
but originally only the Rules.traffic_type field was initialised.
Unfortunately when parsing the protobuf, we need to know the
traffic_type when creating the Table (before the Rules part has been
parsed), so copy the traffic_type value from the Rules into the
GPCTable.

VRVDR-46091