QoS: Added old_config option to Provisioner unit tests
The Provisioner unit tests were only testing the cases where a new QoS
configuration is applied over an empty configuration. The present change
allows an old configuration to be provided to the tests, expanding the
types of tests that can be done. This is needed for a few test cases of
QoS on LAG.

VRVDR-49815