QoS: QoS on LAG code must be executed only when the feature is enabled
The QoS on LAG changes must be effective only on s9500 platform. The
behavior in other platforms must not be affected. A new method has been
implemented that checks the existence of the QoS on LAG feature file.
The Provisioner object checks if the feature is enabled and decides which
QoS configuration objects must be instantiated (QosConfigBondMembers on
s9500 platforms, and QosConfig on all the other platforms).

Fixes: VRVDR-49815