QoS on LAG: Update the LAG notification handling logic
Update the logic that handles the VCI notifications from the LAG component,
using the new class BondMembership. Two QoS configuration objects are
created, one with the previous LAG membership state and the other one
with the new membership state received in the notification. The membership
states are represented by BondMembership objects and passed to the QoS
configuration objects. The Provisioner object compares both QoS
configuration objects and applies the difference onto the dataplane.
Once the notification is processed, the local LAG membership cache is
updated with the new state from the notification.

VRVDR-54214