J2/GPC: Update validation to check not running out of statistics resources
On the J2 there is a maximum number of statistics that is supported for
QoS rules, due to the statistic engine allocated for them.  This adds a
check to ensure the maximum is not exceeded, so configuring too many is
detected at configuration commit time, rather than failing to configure
the statistics on a rule and cause an error log in the journal.

VRVDR-54789