Remove micro-seconds-period from profile and shaper
vyatta-policy-qos-groupings-v1.yang only specifies period not micro-seconds-period.

self._period is initally recieved as a string in milliseconds (eg. '1.01')
It is converted to a float (eg. 1.01)
It is converted to microseconds (eg. 1010.0)
It is converted to an int (eg. 1010)

JIRA: VRVDR-54950
(cherry picked from commit ce72bcc0a9717327ac688bc3106b2d5dd6851313)