GPC-VCI: designation 0 not sent down to dataplane
Designation 0 wasn't being sent down to the dataplane due to a coding
error in FilterAction:add_action.  When the designation was zero the
"if self._designation:" condition would evaluate to false and the
designation would not be added to the action message.

The drop-precedence check was similarly flawed.

Fixes: VRVDR-55098
(cherry picked from commit 40001b3b5f1a53a1271eb9136387e71d5224b4af)