VCI: policer.py didn't handle '1g' bandwidths
The yang pattern statement for the bandwidth-suffix typedef in
vyatta-policy-qos-grouping-v1.yang allows bandwidths of the form
'1k', '2m', '3G', '4ki', '5Mi', '6GI', that is without the trailing
'bps' or 'bit'.  The parse_bandwidth function in policer.py did not
handle these forms of bandwidths and generated the "unsupported operand
type(s)" error.

Part of: VRVDR-55307