Tighten up bandwidth yang pattern
The bandwidth-auto-suffix and bandwidth-suffix typedefs allow bandwidths
of the form '1Bbit', '1Bibps' where the 'B' stands for billions.

However the string is then passed to Rate.pm:parse_rate to convert the
string into a numerical value.  The Rate.pm script only understands SI
prefix symbols, of which 'B' is not a member.  The SI equivalent to 'B'
is 'G' which is fully supported.

Fixes: VRVDR-55307