Checks traffic class does not have more than 8 queues
Remove perl check and replace with python check.

This could not be a yang must statement because if the must statement is
on the traffic class node then the error will be printed 8 times. It can
not be on the profile-opts node because then current() does not refer to
a single traffic-class which means there is no reference traffic-class
to check against in the must statement.

Remove Perl: profile-with-a-traffic-class-with-more-than-eight-queues

VRVDR-48695