VCI: stop double programming of interfaces
Using the "set policy qos platform priority-local-designation" command
meant that each interface in the configuration would be added to the
provisioner object's self._if_updates list by the _check_platform_params
method.

In the next method to be called, _check_interfaces, interfaces could be
added to either the self._if_updates list or the self._if_creates list.

When an interface appears twice on the self._if_updates, or once on the
self._if_updates list and once on the self._if_creates list, then the
command list for this interface would be sent down to the dataplane
twice in quick succession.   This would cause the vyatta-dataplane to
fail to program the QoS configuration because it would find that it had
a configuration already in place when trying to handle the second
attempt.

To prevent this from happening the _check_platform_params method no
longer add interfaces to the self._if_updates list when the local
priority designation needs to be set, but returns a value that is passed
into the _check_interfaces method that tells it if all configured
interfaces need updated.

VRVDR-55714