Remove DRAM check
DRAM fails when the yang file to check has dependent changes in other
repositories that have not yet been merged.

The team have decided because yang changes usually have circular
dependencies the problem can not be solved by merging the PRs
sequentially, starting from the bottom of the dependency chain and
working upwards from there.

This change removes all yang checks. The team are open to adding some
form of check in once we understand the best path forward.

VRVDR-55724