Only run pytest on vyatta_policy_qos_vci
If pytest has no arguments it will look through all the files with test
in their name.
This means it find qos_netconf_test. Which is a system test rather than
a unit test.

This is problematic for 2 reasons.

1. It's dependencies need to be installed otherwise pytest will complain

2. After running dpkg-buildpackage it copies the python files into a new
sub directory.
This leads to pytest finding 2 copies of the files and it complains with
the following:
```
import file mismatch:
imported module 'qos_netconf_test' has this __file__ attribute:
  /home/ag474u/Code/vplane-config-qos/debian/
          vyatta-vplane-qos-tests/usr/share/Qos/tests/qos_netconf_test.py
which is not the same as the test file we want to collect:
  /home/ag474u/Code/vplane-config-qos/tests/qos_netconf_test.py
HINT: remove __pycache__ / .pyc files and/or use a unique basename
for your test file modules
```

This change specifies to pytest which folder to look into to find the
unit tests.

If more unit tests are added in a different location then this will
need to be updated (or removed).

VRVDR-55724