Configure pytest to ignore system tests
Pytest is used as the test runner to invoke the unit tests.
To find the unit tests it looks through the repo finding files with
test in their name. We don't want it looking at the system tests because
it will be unable to parse those files if it does not have the correct
dependencies.

For the debian packaging vyatta-vplane-qos-tests is used to package
system tests. The json files are only used in the unit tests.
So they are not required in the package.

After invoking the debian packaging it copies all the packaged files
to vplane-config-qos/debian/vyatta-policy-qos-vci/usr/lib/python3/
   dist-packages/vyatta_policy_qos_vci/.
This confuses pytest because it ends up seeing the same files twice.
This commit fixes this by telling pytest to also ignore the /debian
directory.

Rename /test to /system_tests to clarify the purpose of the tests.

Remove outdated README for the python tests.

VRVDR-56113