GPC-VCI: don't write an empty protobuf to cstore
After a GPC QoS filter-classification configuration was deleted, the
GPC-VCI daemon would store an empty GPCConfig protobuf in the
vplane-controller's cstore.  This protobuf woud be passed down to the
vyatta-dataplane making it think that the GPC QoS feature was active.
This meant that some unused data-structures were left lying around
inside the vyatta-dataplane.

The fix is to check the number of filter-groups that are part of the
configuration and it the number is zero, use the cstore's "DELETE"
command to delete the pre-existing configuration protobuf, rather than
the "SET" command to overwrite the pre-existing configuration protobuf
with an empty protobuf.

VRVDR-56145