Remove yang backwards compatibility issues
VRVDR-56386

Deviate the must check for a QAX so that it isn't included in the yang.

Check the platform type in the VCI component and if its a QAX check the
wred parameters and adjust so the feature will work.

(cherry picked from commit 070b660d1b7bbee88f28aee0b591a828e3f4b915)